SPIKE: add lock-free restore flag
This is a horrible test-free refactor-free implementation.

Also, we may not even need the lock-free restore flag. No-one is
likely to care about API downtime during a restore, and we probably
don't need to explicitly skip blobstore jobs since the backup data
will be missing from the artifact anyway. So I _think_ the regular
restore process wouldn't mess with the blobstore anyway.

However, if we're going to lean on that, we'll need to test it. For
the initial, experiments, I'm using this flag to explicitly skip
blobstore restore operations anyway.

[#174147726]